Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@8696 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/cintltst/cmsgtst.c
@@ -934,7 +934,7 @@ static void OpenMessageFormatTest(void)
 
     /* Test umsg_applyPattern           */
     status = U_ZERO_ERROR;
-    umsg_applyPattern(f1,pattern,strlen(PAT),NULL,&status);
+    umsg_applyPattern(f1,pattern,(int32_t)strlen(PAT),NULL,&status);
     if(U_FAILURE(status))
     {
         log_err("umsg_applyPattern failed. Error %s \n",u_errorName(status));
